Rewrite types & serialization
