+++ zapisy/apps/theses/assets/theses_app/app_types.ts
@@ -1,4 +1,4 @@
-import { ThesisTypeFilter } from ".";
+import { ThesisTypeFilter } from "./protocol_types";
 
 /**
  * @file Miscellaneous types
